+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Left align the decision data download button
 - Use a multi-line text input for profession registration requirements
 - Enforce length limits on all input fields
+- Add validation to prevent empty and duplicate countries when entering decision data
 
 ## [release-024] - 2022-05-30